requirements.txt
# VKSerbot Unified - Dependencies
# Install: pip install -r requirements.txt
# TIDAK PERLU venv - langsung install ke system Python

# Core (WAJIB)
requests>=2.28.0

# Optional - Instagram features
instagrapi>=2.0.0
pydantic>=2.0.0
Pillow

# Optional - Telegram features
Telethon>=1.42.0
cryptography
pyaes
rsa

# Optional - Other
faker
